+++ CHANGES.md
@@ -3,6 +3,7 @@
 version 0.2.2-dev:
   - Alpha for map polarimetry ploting
   - More direct way to calculate polarimetry angle (no more pccdpack way)
+  - New image registration methods, with footprints
 
 version 0.2.1:
   - Astrometry solving in SimpleCalibPipeline